Update patch set 1
Patch Set 1:

TBH I think this is just creating busywork and code churn.  If we're going to use com.gerritforge as the package, why not just amend the initial commit to do it from the beginning?

But in the end it's your plugin; I'm not going to block it.

Patch-set: 1