Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> TBH I think this is just creating busywork and code churn.  If we're going to use com.gerritforge as the package, why not just amend the initial commit to do it from the beginning?
> 
> But in the end it's your plugin; I'm not going to block it.

GDPR deadline is 25th of May, all those commits are already in master on GitHub anyway and built on CI for deployment on GerritHub.io. I needed to speed this up because of the EU regulations timeline. However, reviews are really appreciated and thanks for raising the point of the plugins' root package.

I believe we should discuss anyway a common policy at the next summit, because at times it is confusing, like the 'com.vmware' of the owners plugin, which more than once raised the doubt in our clients "do I need a VMware license for it?" :-(

Patch-set: 1
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>