Auto-redirect to login when not logged in
When visiting the account self-removal form, make sure
that the current user is in session. If not, redirect to login
and, once logged in, redirect back to the form.

This allows to include the account self-removal form as deeplink
and enable a frictionless flow for users to check their personal
information and remove themselves if required.

Change-Id: I30280d48799a8880b15434eac2dad8583dbd6449